+++ app/page.tsx
@@ -1,5 +1,5 @@
-import DealView from "@/app/(deal)/components/DealView";
 import MainView from "@/app/components/MainView";
+import DealView from "@/app/product/(deal)/components/DealView";
 import PersonalizedProduct from "@/app/product/components/PersonalizedProduct";
 
 type Props = {
